Fix lingering bug in `more_attributes` serializer
Serialization would not happen because setting single values in the
`attributes` hash does not propagate back to the record object. We now
use the regular attribute accessor again.

Also, to automatically migrate records that might still have strings in
their `more_attributes` field, we parse all `more_attributes` strings
with the JSON decoder.

If a `more_attributes` string is invalid, do not allow to save the
record but also don't just discard the invalid string. Since we are
currently exposing the JSON strings to the UI, leave the invalid values
in the record so that the user can easily fix them in the edit form.