Nested form for DependencyConstraints in Ingredient form.
When CREATING or UPDATING an Ingredient, one can now add a
DependencyConstraint through a nested sub form. The user selects the
target Ingredient of the DependencyConstraint, the source_id and
ingredient_id for the DependencyConstraint are inferred by the back-end.

The SHOW action for an Ingredient now lists all DependencyConstraints
that concern the Ingredient itself or any of its children down the
hieararchy tree.