Add CPU constraint
Assuming that more CPUs are always better this constraint is
formulated analogous to the RAM constraint such that instances with
more CPUs also satisfy lower CPU requirements.

The number of CPU cores available is the only metric we consistently get from
each provider. Therefore, we currently only consider the number of cpu cores.