Attempt to fix the ingredient ordering issue #174
The cause of this ordering issues seems to be the different ways `all_leafs` are accessed.
Sometimes, they are sorted by id (ingredient_resource_mapping, remap_more_attributes) and
sometimes they are not sorted by id (generate_ingredients_data).

Currently, this issue occurs only in combination with the production data because
the ordering of the ingredient ids in production is different.