Add job id list to track jobs per ingredient
Notice that the creation of a job (via `perform_later`) and the
job id tracking (via `add_job`) need to happen in the same transaction.
Otherwise, jobs could already be processed and the job id is added afterwards
resulting in an excess job id. In this case, recommendations would never be marked
as completed for such ingredients.